requirements.txt
asgiref==3.2.10
boto==2.49.0
boto3==1.15.13
botocore==1.18.13
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.1.1
Django==3.1.1
django-cors-headers==3.5.0
django-rest-knox==4.1.0
django-rest-swagger==2.2.0
django-storages==1.10.1
djangorestframework==3.11.1
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
mysqlclient==2.0.1
numpy==1.19.2
openapi-codec==1.3.2
pandas==1.1.2
Pillow==7.2.0
psycopg2==2.8.6
pycparser==2.20
PyMySQL==0.10.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
s3transfer==0.3.3
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.10
